Zero values should be omitted
If a struct has a lot of zero values we'll currently send a lot of data about zeros.  That's not wonderfully useful.

But if we fix it then we have issues around slices of pointers to structs with all zero values versus nil. We won't be able to distinguish them,